+++ Assets/Plugins/UnityFx.Outline/Runtime/Scripts/OutlineLayerCollection.cs
@@ -20,14 +20,47 @@ namespace UnityFx.Outline
 	{
 		#region data
 
+		private class OutlineLayerComparer : IComparer<OutlineLayer>
+		{
+			public int Compare(OutlineLayer x, OutlineLayer y)
+			{
+				return x.Priority - y.Priority;
+			}
+		}
+
 		[SerializeField, HideInInspector]
 		private List<OutlineLayer> _layers = new List<OutlineLayer>();
 
+		private List<OutlineLayer> _sortedLayers = new List<OutlineLayer>();
+		private OutlineLayerComparer _sortComparer = new OutlineLayerComparer();
+		private bool _orderChanged = true;
 		private bool _changed = true;
 
 		#endregion
 
 		#region interface
+
+		/// <summary>
+		/// Gets layers ordered by <see cref="OutlineLayer.Priority"/>.
+		/// </summary>
+		public OutlineLayer[] SortedLayers
+		{
+			get
+			{
+				UpdateSortedLayersIdNeeded();
+				return _sortedLayers.ToArray();
+			}
+		}
+
+		#endregion
+
+		#region internals
+
+		internal void SetOrderChanged()
+		{
+			_orderChanged = true;
+			_changed = true;
+		}
 
 		internal void Reset()
 		{
@@ -45,6 +78,16 @@ namespace UnityFx.Outline
 			}
 		}
 
+		internal void Render(OutlineRenderer renderer, OutlineResources resources)
+		{
+			UpdateSortedLayersIdNeeded();
+
+			foreach (var layer in _sortedLayers)
+			{
+				layer.Render(renderer, resources);
+			}
+		}
+
 		#endregion
 
 		#region ScriptableObject
@@ -55,6 +98,8 @@ namespace UnityFx.Outline
 			{
 				layer.SetCollection(this);
 			}
+
+			_orderChanged = true;
 		}
 
 		#endregion
@@ -86,6 +131,8 @@ namespace UnityFx.Outline
 
 					_layers[layerIndex].SetCollection(null);
 					_layers[layerIndex] = value;
+
+					_orderChanged = true;
 					_changed = true;
 				}
 			}
@@ -115,6 +162,8 @@ namespace UnityFx.Outline
 				layer.SetCollection(this);
 
 				_layers.Insert(index, layer);
+
+				_orderChanged = true;
 				_changed = true;
 			}
 		}
@@ -126,6 +175,8 @@ namespace UnityFx.Outline
 			{
 				_layers[index].SetCollection(null);
 				_layers.RemoveAt(index);
+
+				_orderChanged = true;
 				_changed = true;
 			}
 		}
@@ -165,6 +216,8 @@ namespace UnityFx.Outline
 				layer.SetCollection(this);
 
 				_layers.Add(layer);
+
+				_orderChanged = true;
 				_changed = true;
 			}
 		}
@@ -176,7 +229,9 @@ namespace UnityFx.Outline
 			{
 				layer.SetCollection(null);
 
+				_sortedLayers.Remove(layer);
 				_changed = true;
+
 				return true;
 			}
 
@@ -194,6 +249,7 @@ namespace UnityFx.Outline
 				}
 
 				_layers.Clear();
+				_sortedLayers.Clear();
 				_changed = true;
 			}
 		}
@@ -270,6 +326,18 @@ namespace UnityFx.Outline
 		#endregion
 
 		#region implementation
+
+		private void UpdateSortedLayersIdNeeded()
+		{
+			if (_orderChanged)
+			{
+				_sortedLayers.Clear();
+				_sortedLayers.AddRange(_layers);
+				_sortedLayers.Sort(_sortComparer);
+				_orderChanged = false;
+			}
+		}
+
 		#endregion
 	}
 }