+++ CHANGELOG.md
@@ -3,9 +3,27 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/); this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.6.0] - 2019.09.26
+
+Quality of life improvements.
+
+### Added
+- Added `OutlineLayer.Enabled`.
+- Added `OutlineLayer.Name`.
+- Added possibility to change render order of layers via `OutlineLayer.Priority`.
+- Added possibility to edit renderers of an `OutlineLayer`.
+- Added possibility to alter `CameraEvent` used to render `OutlineEffect`.
+- Added more info to the `OutlineLayer` preview inspector.
+
+### Changed
+- `IOutilneSettings` setters now throw if overriden.
+
+### Fixed
+- Fixed `OutlineLayer.Add` not filtering renderers by the mask passed.
+
 ## [0.5.0] - 2019.09.09
 
-Features editor UI and unit tests.
+Editor UI improvements and unit tests.
 
 ### Added
 - Added `OutlineSettings`, that can be shared between dfferent `OutlineLayer` and `OutlineBehaviour` instances.
@@ -18,7 +36,7 @@ Features editor UI and unit tests.
 
 ## [0.4.0] - 2019.08.31
 
-Features blurred otulines.
+Blurred outlines.
 
 ### Added
 - Added Gauss blurring to outlines.